Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -5,7 +5,7 @@
 void main(){
 int a,b;
 scanf("%d %d",&a,&b);
-double n=pow(a ,1/2);
-printf("%s", (int)n==b ?"Yes":"No");
+int n=pow(a ,1/2);
+printf("%s", n==b ?"Yes":"No");
 
 }
